+++ tests/test_groups.py
@@ -1,8 +1,11 @@
 # coding: utf-8
 from __future__ import unicode_literals
 
-import pytest
 import time
+
+import pytest
+
+from nuxeo.compat import text
 
 
 @pytest.fixture(scope='function')
@@ -37,7 +40,7 @@ def test_fetch_unknown_group(server):
 
 
 def test_update_group(server, group):
-    grouplabel = str(int(round(time.time() * 1000)))
+    grouplabel = text(int(round(time.time() * 1000)))
     group.grouplabel = grouplabel
     group.save()
     group = server.groups().fetch('plops')